Moved path normalization
